refactor: change structure of client`s package in scrapper
