--- scrapper/src/main/java/edu/java/client/StackOverflowClient.java
@@ -1,7 +1,0 @@
-package edu.java.client;
-
-import edu.java.response.QuestionResponse;
-
-public interface StackOverflowClient {
-    QuestionResponse fetchQuestion(Long questionId);
-}
