+++ scrapper/src/main/java/edu/java/client/github/GithubClient.java
@@ -1,6 +1,6 @@
-package edu.java.client;
+package edu.java.client.github;
 
-import edu.java.response.RepositoryResponse;
+import edu.java.client.github.dto.RepositoryResponse;
 
 public interface GithubClient {
     RepositoryResponse fetchRepository(String ownerUserName, String repositoryName);
