+++ scrapper/src/main/java/edu/java/client/stackoverflow/StackOverflowClient.java
@@ -1,0 +1,7 @@
+package edu.java.client.stackoverflow;
+
+import edu.java.client.stackoverflow.dto.QuestionResponse;
+
+public interface StackOverflowClient {
+    QuestionResponse fetchQuestion(Long questionId);
+}
