+++ scrapper/src/main/java/edu/java/client/stackoverflow/dto/QuestionResponse.java
@@ -1,4 +1,4 @@
-package edu.java.response;
+package edu.java.client.stackoverflow.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
